requirements.txt
Django==1.7.1
Markdown==2.4.1
PyYAML==3.11
anyjson==0.3.3
click==3.3
django-ckeditor==4.4.6
django-multiselectfield==0.1.3
djangorestframework==2.4.0
pep8==1.5.7
pytz==2014.4
six==1.7.3
wsgiref==0.1.2
Pillow==2.6.1
django-sendfile==0.3.6
autopep8
django-mathfilters==0.3.0
django-grappelli
django-nose
django-extensions
django-debug-toolbar
django-nocaptcha-recaptcha==0.0.18
django-recaptcha==1.0.4
